fix: stack histograms from variable number of columns
The data in FOO has different numbers of columns on different lines.
The reported error case is
	set style data histograms
	set style histogram rowstacked
works:
	plot for [i=1:*] FOO using i
fails:
	plot for [i=1:*] FOO using (column(i))

The difference is that when i > actual number of columns
column(i) returns Nan and DF_UNDEFINED to getdata()
the bare column number returns garbage.

This is a partial fix.  It catches the stacked histogram case.
I think a larger-scale fix to handle all cases of (i > df_no_cols)
is needed, but that's a bigger change than I want to make for 6.0rc3.