EXPERIMENTAL "set mouse multiplot" enables use of remultiplot while mousing
Auto-substitution of remultiplot for replot when the most recent
plot command was a multiplot is already handled when "replot" is
issued from the command line.

The intent is to make the analogous test and substitution also while
mousing, but for the initial release of 6.0 let's be conservative.

Rather than making remultiplot during mousing active by default,
"set mouse multiplot" enables the analogous substitution
of remultiplot for replot/refresh when the request is issued through
do_string_replot().

This is marked EXPERIMENTAL, and defaults to "nomultiplot",  because it is
relatively untested and because additional improvements may come later.