revert introduction of "set mouse multiplot"
This command was added in -rc3 because the multiplot mousing code was
very new and basically untested.  "set mouse nomultiplot" disabled it.
But now (hah) the code is cleaned up and we can unconditionally enable
multiplot mousing in the full 6.0 release.

reverts most of  0930058b and part of 0f516c69