"plot with polygons" now guarantees that each polygon has a closed border
The polygon code piggy-backs on the filledcurve code, but filled curves
do not necessarily have closed borders.  Now we distinguish "with polygons"
as a separate case and check each one to see if it requires drawing an extra
line to connect the last and first points.

This also fixes the variable color handling for polygons.