last-minute fix to prevent memory corruption if "splot ... zsort"
"zsort" is not currently used in 3D plots, but df_open was trying
to parse it as a filter option for both 2D and 3D data files.
Since the plot_filter header field is not in the shared 2D/3D header
block, this overwrites (struct contours *)contours instead. Boom.

Must be fixed properly for 6.0.1 and 6.1 !